# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items with optional filters
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: belowMin
          in: query
          schema: { type: boolean }
        - name: nearExpiry
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated list of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
    put:
      operationId: updateItem
      tags: [items]
      summary: Update item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Deactivate item (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deactivated
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment with optional filters
      parameters:
        - name: condition
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated list of equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListResponse"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create new equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
...
  /transactions/out:
    post:
      operationId: createOutTransaction
      tags: [transactions]
      summary: Record an outbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /recipients:
    get:
      operationId: listRecipients
      tags: [recipients]
      summary: List recipient organizations
      responses:
        "200":
          description: List of recipients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipient"
    post:
      operationId: createRecipient
      tags: [recipients]
      summary: Create a recipient organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipientInput"
      responses:
        "201":
          description: Created recipient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
  /exit-reasons:
    get:
      operationId: listExitReasons
      tags: [exit-reasons]
      summary: List exit reasons
      responses:
        "200":
          description: List of exit reasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExitReason"
    post:
      operationId: createExitReason
      tags: [exit-reasons]
      summary: Create an exit reason
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExitReasonInput"
      responses:
        "201":
          description: Created exit reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExitReason"
...
components:
  schemas:
...
    LoginCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, fullName, role]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, warehouse_manager, viewer]
...
    Item:
      type: object
      required: [id, name, unit, itemType, currentStock, minStock, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: ["string", "null"]
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        itemType:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minStock:
          type: integer
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ItemInput:
      type: object
      required: [name, unit, itemType]
      properties:
        code:
          type: ["string", "null"]
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        itemType:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minStock:
          type: integer
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ItemUpdate:
      type: object
      properties:
        code:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        itemType:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        minStock:
          type: ["integer", "null"]
        expiryDate:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    ItemListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Equipment:
      type: object
      required: [id, name, condition, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
          enum: [good, maintenance, broken, consumed, needs_inspection]
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EquipmentInput:
      type: object
      required: [name, condition]
      properties:
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    EquipmentListResponse:
      type: object
      required: [equipment, total, page, limit]
      properties:
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/Equipment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Transaction:
      type: object
      required: [id, type, itemType, documentNumber, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, init]
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        itemUnit:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientName:
          type: ["string", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        exitReason:
          type: ["string", "null"]
        documentNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OutTransactionInput:
      type: object
      required: [itemType]
      properties:
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    Recipient:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
    RecipientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        notes:
          type: ["string", "null"]
    ExitReason:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    ExitReasonInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
...