# lib/api-zod/src/generated/types/equipmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ambulance & Emergency Warehouse Management System API — Damascus
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentInput {
  name: string;
  /** @nullable */
  equipmentType?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  condition: string;
  /** @nullable */
  manufactureYear?: number | null;
  /** @nullable */
  originCountry?: string | null;
  /** @nullable */
  currentHolder?: string | null;
  /** @nullable */
  notes?: string | null;
}