# lib/api-zod/src/generated/types/outTransactionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ambulance & Emergency Warehouse Management System API — Damascus
 * OpenAPI spec version: 0.1.0
 */
import type { OutTransactionInputItemType } from './outTransactionInputItemType';

export interface OutTransactionInput {
  itemType: OutTransactionInputItemType;
  /** @nullable */
  itemId?: number | null;
  /** @nullable */
  equipmentId?: number | null;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  recipientId?: number | null;
  /** @nullable */
  recipientPerson?: string | null;
  /** @nullable */
  exitReasonId?: number | null;
  /** @nullable */
  notes?: string | null;
}