# lib/api-zod/src/generated/types/recipient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ambulance & Emergency Warehouse Management System API — Damascus
 * OpenAPI spec version: 0.1.0
 */

export interface Recipient {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
}